Add Priority Feature
Priority tracking is not currently implemented in Elp.

A priority tracking feature can help users better prioritise and
keep track of tasks.

Adding a priority field to the tasks ToDo, Event and Deadline
allows for the priority to be displayed as well as the other
relevant fields to be displayed.

Let's add the possible priorities to a Priority Enum class and
adjust the class fields of the tasks and Storage class.

Adjusting the composition of the tasks requires additional
refactoring for less code duplication and adjustments to the
Storage class for storing/loading contents into the save file in
the hard drive of the user.