Add proper padding and sizes of GUI elements
Padding is added to InputContainer and DialogContainer
for a cleaner look.

Sizes of GUI elements that used to be magic literals were
properly abstracted out as class attributes of the various
GUI elements.

Font Size of text in DialogBoxes is fixed at 12.

Response messages were also updated to no longer include
tabspaces at the beginning for standardization of response formats.