Improve Code Quality
Some classes and methods lack javaDoc headers. In particular, code for the different enums in Parser class are deeply nested and hard to read and understand.

Fixing the code quality makes the code easier to read and understand for users or other future developers who may reference the code.

Let's
* add javadoc comments and headers for all class and methods
* make sure code is not too long
* create new Command classes to remove deep nesting in Parser class