Sort tasks whenever the full task list is requested
Tasks are not sorted.

Sorted tasks is more user-friendly, as Events, Deadlines and Todos will be displayed in sequence. Events and Deadlines
with earlier durations and deadlines are also displayed first.

Task implementing Comparable allows us to give an order to the various types of Tasks. For example, Deadlines can come before Events
followed by Todos. Within Deadlines and Events, they can also be sorted by their deadlines and durations.

The TaskList is only sorted when the user gives the list command so that it is more efficient.