Improve code quality
The code quality is not so good, since to long method is detected in the
code, and readability of the code needs improvement.

Let's shorten the length of parseCommand method in Parser class
Let's add a command package to handle different types of command
Let's add more helper methods to Parser class

Using better OOP and SLAP mechanism can make the code cleaner and more
understandable to readers.