Add assertions
Assertions are needed to define assumptions about the program state,
assetions failures help to indicate if there is a bug in the code.

Let's add assertions to Storage.java to check that tasks are not null,
and at BobTheBot.java to check that bot will only enter exit command
when "bye" is entered by the user.