Add assertions in the code.
No assert statements in the code.

It needs to change so that important assumptions that should hold at
various points in the code is documented.

Add assert statements to Task.java for getDescription and getIndex
methods, and to TaskList.java for the addEvent, addToDo, addDeadline
methods.

Using assertions in these areas ensure the assumption holds for each
method, that description for tasks are never empty, and there is a
task being added into the TaskList every time the add function is
invoked.

Assertions are not only simple, but their impact on performance is
considered low and worth the additional safety they provide.