Add tag to tasks and update storage methods
Add tag field to Task class and added new constructors to Task and its subclasses to make the creation of these tasks more flexible.

Storage will use constructors that require a tag to be passed as tags will always be stored in the storage file. If it is empty, it will be stored as null.

The commands will use contructors that do not require a tag as they will not be able to assign a tag to a new task.