Added new archive feature to dukebot
Currently, dukebot was is only able to delete tasks without saving.

This caused an issue to users who want to remove tasks from main list without deleting it full.

Now with the archive command, users can archive the whole list or specific tasks from the main list to the archive list for reference in the future.