Add UndoCommand to chatbot
The user is unable to conveniently undo his mistake after inputting a
command that changes the state of the tasklist.

In cases where the tasklist starts to get very long, the user may have a
hard time looking for the last change he did to manually revert the
change himself.

Let's add a new UndoCommand to allow the user to revert the tasklist to
the state before the previous change occurred.