Storage: create directory for save file if it does not exist
There is no check for if the directory containing save file exists. This leads to
potential bugs if the directory is not present and we attempt to access the save file.

Users are not expected nor instructued to create this directory. Thus, we need to create
the directory if it does not exist.

Let's
* add check if directory exists
* create directory if it does not exist