Storage: create directory or file if nonexistent when saving
There is no check if the save file or directory exists when saving. There
is such a check for loading a file.

Successfully loading the save file at the start does not ensure the file
will remain when saving to it later. Thus, we need to perform the same
check when saving.

Let's,
* abstract out the logic of creating the new save file if nonexistent
* call the validation method before saving and loading