Add parser method and command class for each user input
Improve code quality in 3 ways:
1. Refactor code to have different methods for parsing
   different user input.
2. Add commands to handle the result of the use input on
   the ToDoList.

Having different methods to handle the parsing and output of
user input helps to reduce method length and leads to cleaner
code quality.

We should have a parser method to handle each input, and a
command class which handles how the user input interacts with
the ToDoList class.