C-Update. adds update functionality
After creating tasks, you cannot update (ie: edit) them directly.

To edit a task, a user would have to first delete the task, then
create a new task with the updated details.

Let's simplify this process by allowing the user to directly update
an existing task. The update function will edit the task details
according to user's inputted details. It will not delete and create
a new task so that the order of the task in the task list remains.

However, the command to update a task might be a little troublesome
since the user would have to edit all the given fields (eg: even
if the user only wants to edit the end date of an event, the user
would have to re-enter the event details and start time as well).
Future commits may want to look into improving this aspect.