Add assertions to code
Assertions are not existent in the code to verify bugs.

As such, certain bugs might not be detected in code.

Use assert feature to document important assumptions that should hold
at various points in the code.

This is to prevent major bugs in the code such as description of
Tasks being null.