Improve code quality
Code quality for project improved in accordance to Java standards and
conventions.

Coding in accordance to Java standards and conventions makes the project
more readable to other developers to work on and debug.

Let's edit the package names, variable names and comments such that they
are now in accordance to convention.