Add assertions to Duke
There are no assertions throughout the application.

It is recommended that assertions are used liberally in the code. Their impact on performance is considered low and worth the additional safety they provide.

Adding assertions to methods in duke related to task lists and storage allows us to ensure an unusual condition created by the user raises an exception.

In methods that require tasks list, we assert that task list is not null. In methods that require the storage, we assert that the storage is not null. In methods that require at least one task, we assert that the task list is not empty.  The same concept is applied to file and inputs.