Add assertions
The code assumes certain properties during runtime. This may cause bugs
in the program if the properties are not set correctly during runtime.

Add assertions to ensure that the assumptions made in the program during
runtime are valid.

Assert statements are added in the following classes which are where
crucial assumptions are made,
* Storage
* TaskList