Improve coding standards
There are some violations in coding standards throughout the application.

We should practice writing code of high quality.
Given how the world is becoming increasingly dependent on software,
poor quality code is something no one can afford to tolerate.

There are two main improvements made in this commit.
*Avoid Complicated Expressions
*Make the Happy Path Prominent

This is done to increase readability of code which will lead to increase
efficiency in debugging and extension in the future.

Javadocs were added to some methods thoughout the application as well.