Update Duke.java update Storage.java update Parser.java
Method comments were a little confusing in Duke.java and Parser.java

This makes it difficult to read and understand their usages.

Improving comment readability allows others to understand the code
more easily and improve overall code quality.

Parser's parse method was long and complicated.

This makes it difficult to understand the logic of the code

Efforts were made to improve readability, but method remains a litte
long  as the current implementation requires a drastic change
in the base code of most other classes. Readability and code logic
was improved as much as possible.