Add Help Command
There are no commands to show the user what commands there are.
The only help occurs when the user types a wrong command.

We should be explicit in allowing the user know they can get help in using Duke.
We must implement a help command so users can get help.