Fix bug where single word command results in exception
Currently, the command is obtained from user input by looking for the first " " character. However, if the input is a single word command then a -1 is returned, causing string out of bounds error.

To fix this, a check was introduced to prevent substring generation if the input is a single word command.