Improve code quality in InputParser
Previously, the parse() method is lengthy. The body is divided into two main
responsibilities - separating the input into the command keyword and the body,
and then parsing them to return a Command. Code blocks are extracted into two
private methods to improve the readability of the code and adhere to a single level\
of abstraction within parse().