Updated Duke.java
When GUI was opened with FXML, initial items in file was not being updated.
When running in CLI, this problem did not exist. This was due to the fact
that storage.openFile() and storage.listInit() was only called in Duke.run()
method, which was not called when the GUI was run.

Methods mentioned were added to the constructor to ensure that the
tasks saved in the hard disk was properly added to the list when run
in the GUI.

Ensured that the program runs as intended both when using CLI and
when using the GUI.