Refactor code to improve code quality
Modified program so that the execution of commands (e.g. adding task to
the task list and also writing to storage) is handled by Duke.java
instead of Command.java.

Also shortened methods in Command.java and Storage.java to improve
readability, and changed names of Boolean variables in accordance with
the CS2103T naming conventions.