Updated Parser.java Updated StringReplacer.java
Users did not have an option to edit their tasks without first
deleting them and re-adding them back.

An edit function was added to allow editing of both deadline and
event tasks (tasks with date and time). This allows users to
update the date and time of such events in the case of a change
in their schedule without having to first delete the task.

Update can be called with the following command:
update <task number> /<date and time>