Debug the saving and loading of data
Currently, the saving and loading of data does not work as the saveToFile method saves an empty text file.

Debug the code by allowing the Parser to load the text file when initiated.

This allows the program to load the previously saved data easily and create a new TaskList when there is no data.