Add Reminder Function
Duke is unable to provide a reminder function that can remind the user
of upcoming events or deadlines in the coming week.

Reminder function should be added so that user can easily know what
are the upcoming events and deadlines.

Let's have a reminder function that can allow the user to know about
the upcoming deadlines and events within the next week.

Typing in "Reminder" returns the deadlines and events for the next
7 days. E.g. Typing "Reminder" on 12 September 2022 will return the
deadlines and events from 12 September 2022 (Exclusive) to 19 September
2022 (Exclusive).