Add Assertions to Duke
Assertions allow for easy debugging when we plan to extend our codebase
in the future iterations

Assertions are added to constructors to make sure Tasks object and
String objects are not null.

Assertions are added to TaskList methods to ensure all index are
non-negative and smaller than size of taskList.