Add assertions
Use assertions to verify the correctness of various Task and Duke classes.

All tasks should have a description.
The getResponse method in Duke should never get an empty input.

Add assertions to Task constructors to ensure the description is not empty.
Add assertion to getResponse method in Duke to ensure the input is not empty.

This will ensure the classes behave as expected during runtime.