Add assertions
The Deadline and Event constructors make certain assumptions about the format of the parameters.

In the future, changes to the Parser class may make the Deadline and Event classes behave unexpectedly.

Let's use assertions to make sure the parameters are in the correct format.

Using assertions allows us to document important assumptions that should hold true for the parameters in the constructors.