Refactors Code for Single Input Parsing
Before this commit, the code was set up in a way that requires multi staged inputs, such as first entering into a task stage, then executing commands in the task stage. As per the suggestions in the guidelines, this is now changed to single line input and will now require execution in one line only.

For example: task add todo do laundry

Some exceptions are reserved for confirmation commands, but need not be answered and can be ignore.

For example: task delete all
> Are you sure you want to delete all tasks?
One can choose to say yes, no, or ignore this prompt altogether and issue another command, on which the previous instruction will be ignored and the new command will be executed instead.

Updates to the GUI added: Dark Mode