Improve Code Quality
All display messages are written in the execute() methods of
each individual command.

Adding all these display messages to the Ui class can help in
modifying these messages in the future and keep the display
messages separate from the logic of the code.

Let's add all display messages to the Ui class to keep them
separate from the actual logic of the code.