Improve code quality
Moved all the boolean conditions to check error into variables with a clear name on the functions they serve and deleted all the unnecessary code.

These allow the function of the code to be more obvious to the readers and avoid confusing the readers on the purpose of the unused code.