Add duplicate detection feature to Duke.
Duplicate tasks can be added into the Duke task list.

This may be a mistake by the user and may cause confusion.

Added the methods isDuplicate to check for duplicates, modified
the add method and added a method into UI class to implement
the feature to prevent duplicate tasks when user tries to add duplicates.