Create a method to generate display messages
The execute method in each command class before returning the execution
message passes it to the returnTask method of the UI class which returns
back the message without making any modifications to it.

Creating a method in the UI class to generate the execution message
based on the type of the command being executed.

This improves the code quality by removing redundant method calls and
avoiding code repitition.