Move exit logic to within ExitCommand
The exit logic is weirdly wired through MainWindow which should not be
responsible for the effects of a Command execution.

Let's move the exit logic using TimerTask into ExitCommand to be
executed. This follows all other type of Commands' behaviours.

A minor note: As running Timer is asynchronous, we are able to achieve
this effect without much issues.