Add more assertions
In rewriteFile function there is a redundant else if check to check
whether the task t is an instance of Deadline or not.

Replacing else if check with else as if the task is not an instance of
Todo or Event then it is compelled to be an instance of Deadline.

Add an assertion within the else block to assert that task t is an
instance of Deadline.