Abstract out the bye command into CommandByeHandler.java.
Previously, the bye command is checked at the user input level.
This is not ideal because the implementation is not extensible.

By capitalizing on the previously built infrastructure of CommandHandler,
I abstracted out the bye command into its own class to have better
extensibility and overall cleanliness of the code.