Fix bugs and improve error handling.
Fixed several off-by-one bugs in MarkCommand, UnmarkCommand, DeleteCommand.

Created independent classes for exceptions that are thrown more than once for neater code.

Changed implementation of FindCommand to use FoundList instead, and return a list of items found that retain the original indices of the items as on the main TaskList.