Add functionality to track expenses.
Expense tracking is now available. Expenses are stored in the
same data file as the Tasks to keep data in a single file. Expense and
ExpenseList classes are added for the Expense tracking functionality.
Storage class is modified to load and store Expenses, Parser is modified
to be able to parse commands related to Expenses. Bye command now
properly terminates Greg.