Parser class: refactor code within parse function
The parse() method is long.

The length of the code in the parse() method negatively impacts
readability, making it difficult to identify the logical flow of
the code. This increases the possibility of errors in the code
going unnoticed.

Refactoring code into smaller, more focused methods allow us to
know where to look when we decide to change certain parts of how
the Parser class checks for different commands.

Let's refactor the command-checking logic into their own methods.

Each command requires multiple lines of code to check for validity.
Refactoring the code into their own methods declutters the parse()
method, and allows for easier finding and changing of checking logic
for the respective commands in the future.