Abstract out shared functions of Event and Deadline
Event and Deadline classes share time-based parameters and methods.

The common parameters and behavior cause code duplication.

Extracting the common behavior into the ListObject superclass
makes the codebase adhere to the DRY principle and scalable in future
to introduce other subclasses dealing with time associated with tasks.

Let's include the time field in the superclass ListObject and
put time reading and modification methods into the superclass.

Using such inheritance is preferable as it makes the codebase
more scalable should we wish to introduce new subclasses
that have a time field associated with the task.