Add resize functionality and more visual improvements to GUI
The GUI is unable to be resized and the colour palette is quite bland,
and does not fit the personality of the bot.

Adding more the resize functionality will make the bot more versatile
and adaptable to different display sizes. Adding a blue/pink colour
palette as well as rounded dialog box corners creates a more friendly
user experience.