Add sort command
Sort command is null safe to handle todo tasks.

A sort is necessary for users to sort tasks chronologically.

Using a custom comparator class for tasks is preferable because other sorting parameters such as description can be added later.

Refer to this S/O discussion on dealing with multiple comparators as a chain
https://codereview.stackexchange.com/questions/229497/handle-null-in-comparator-class