Add Find functionality
Currently, there is no way to find tasks that a user is looking for.

This makes filtering down tasks or tracking tasks difficult for users, especially in the future when they may have multiple tasks saved.

Let's add a find functionality that will filter their tasks based on the search keyword that they provide.

This helps users to easily find tasks that contain their keywords for easier accessibility and filtering.