Add more exceptions and help command
Introduced a new InvalidInputException class to handle invalid
inputs from user. More errors are now handled with exceptions.
More problematic user inputs are now being handled, such as inputting
Strings or blanks for positions, or if the user does not input certain
fields required for the command. A new help command is added to
provide the user with a list of available commands.