Create user guide and bug fixes
No user guide currently exists for the Elp chatbot.

It may be difficult to understand how to properly utilise the
chatbot without knowing the comments and syntax required.

Let's add a user guide for the project as a website to allow for
users to learn how to use the chatbot.

A github website is a clean and easy way to implement this, and can
be found at the URL: https://marclzh.github.io/ip/.