Added GUI with JavaFX
Currently, Duke does not have a GUI for interactive user experiences.

This severely hampers the usability and appeal for laymen users who do not appreciate the CLI interface.

Let's add a GUI here to allow Duke to run as a chatbot-like application.

This would help aid the user experience and improve the readability of Duke.