Add assertions
Code has no assertions.

Bugs in the code may not be detected and terminated early, which is not
safe.

Let's add some assertions to check assumptions about the program state.

Assertions are used as these checks are performed within functional
code. The runtime can verify them early which is important as assertion
failures indicates a possible bug.