Enhance mark, unmark and delete with mass operations
The commands for mark, unmark and delete only supports operating
on one task at a time. However, upon usage, the user may need to
work on multiple tasks at once, for instance marking all the tasks
as done, or deleting a few tasks as done.

Mark, unmark and delete now supports mass operations, as the
keywords "all" or a range of values (e.g. 1-3) can be passed in
as the arguments for the commands.