Convert images to png, disable resizing
Images could not be loaded on Linux,
possibly because they are in jpg format.
Window could be resized, but the proportion of components is incorrect.

Converting image format from jpg to png could possibly solve the issue.
Disable resizing of window provides a quick fix
of window resizing problem.

Let's
* convert image format to png
* disable window resizing