A-Assertions: add assert statements to the code
Currently assumptions are made on invariants as the code logic
is processed.

Assumptions may be made about internal invariants, etc., but these
may not hold in unforeseen circumstances, which will cause the code
and logic to break, and potentially cause irreversible damage.

Let's include assert statements into the code logic.

Assert statements allow us to liberally ensure that these
assumptions are hold as the code is processed, ensuring that
unforeseen circumstances are accounted for,
safeguarding the program and data from damage.

Assert statements are preferred to other methods, as the runtime
will terminate the program immediately, safeguarding the program
from continuing.

Refer to this S/O discussion on why assert is preferred
"https://stackoverflow.com/questions/11468899/any-benefit-
of-using-assert-instead-of-using-a-simple-if"