Command class: empty A-Assertions commit
Assertions are already being used in the Command class.

This is already the case since commit
796ab22c907149a120fe0833bc0db6bcb5f49977.

Committing an empty commit allows for the history to diverge from the
main branch. That way, a PR can be created for this branch. The PR is
needed as it is one of the requirements for this module.

In addition, this full commit description is also necessary to satisfy
another requirement for full commit messages in this module.

Let's create an empty commit to fulfil these requirements.