Span class: empty A-Streams commit
Streams are already being used in the Span class.

This is already the case since commit
4de4b19d08a609069b5d8206cfaba4960de1a66b.

Committing this empty commit allows for the history to diverge from the
main branch. That way, a PR can be created for this branch. The PR is
needed as part of the requirements for this module.

In addition, this full commit description is also necessary to satisfy
another requirement for this module, which are having some commits with
full descriptions.

Let's create an empty commit to fulfil these requirements.