TaskManager class: empty A-Streams commit
Streams are already being used in the TaskManager class.

This is already the case since commit
a94055ad755806a79832f5718c5addffe66feda7.

Committing an empty commit allows for the history to diverge from the
main branch. That way, a PR can be created for this branch, which
fulfils one of the requirements for this module.

In addition, this full commit descripton also satisfies another module
requirement which are having some commits with full commit descriptions.

Let's create an empty commit to satisfy these requirements.