addTask: extract logic into 3 separate methods
Currently, the addTask fully handles the "parsing" of the user
input to choose the appropriate task to be added for the addTask
method.

This can complicate readability of the code, and create confusion.
The parser method should fully handle which type of task should be
added.

Extracting the logic into 3 separate methods enhances the code
quality and readability, as well as helping to reduce arrowheads in
the code, complying with the SLAP principle.

This can set the code design up for further changes to be made in
its appropriate method, allowing for easier extension of features
and changes in the future.