Add assertions
Parts of code rely on certain arguments being exclusively one or two
values.

The arguments being inputted could potentially be outside the range of
intended arguments.

Use assertions to ensure that only the intended arguments are inputted.

Let's add assertions to ensure only intended arguments are the ones
that were inputted to reduce the number of bugs.