Declare filepath in storage
filePath is declared in duke and passed as a parameter for
storage.

Duke class need not know the filePath hence it is unnecessary to declare it.

FilePath is now declared in storage as a final string constant.

FilePath is unchanged and only used in methods in storage hence only storage
needs to know the exact filePath.