Classify classes and use switch case
All classes are under one duke package and parser uses if-else statements which
has more than 3 indentations. One additional change is that  I have used an * for
an import statement.

I have changed the if-else to switch case in Parser class, classified the classed
according to they functions and listed out the import statements instead of using *

To:
- switch-case: make code more readable as it has less indentation
- classify: easier to look for classes to edit and debug
- state imports: avoid importing unused apis