Remove code duplication by adding to parent class
The getStorageDescription method duplicates code 3 times in the ToDo,
Deadline and Event class.

Deadline and Event classes both have a LocalTime variable in it.

Declaring a LocalTime variable in the TimedTask class as well as
implementing the getStorageDescription method in both Task and TimedTask
can reduce the amount of code duplication.

Let's create a LocalTime variable in the TimedTask class as well as
implement a getStorageDescription method in both the Task and TimedTask
classes.

This allows us to initialize less variables in Deadline and Event as
well as use the implementation of getStorageDescription in the parent
classes to implement the getStorageDescription method in ToDo, Deadline
and Event.