Improve code quality
There are some unclear conditionals and no default branch for if-else
statement. There are also long methods with repeated code doing similar
things.

This makes the code difficult to read. It is also prone to error.

Let's
* Redefine conditionals with a name
* Add a default branch and throw an exception if the default branch is
  reached, as it is not supposed to
* Extract repeated code out into methods

Giving boolean conditionals a name makes its purpose more easily
understood. Choosing to still add a default branch even if it is
supposed to be unreachable helps to ensure that the code is written
correctly, and errors should there be an accidentally misuse be easily
detected. Extracting repeated code into methods makes it easy to reuse
the same piece of code for similar purposes.