A-Assertions: add assert method and file checker to storage.
The loadItems method has no check to ensure the tasks file is present.

The loadItems method cannot function without a task file present.

A function to create the file if it is not there and an assert method
to ensure the file is there will be added.

A function and assert are both used as assert cannot be used to carry
out tasks and should be used just as a check