Examines the code and refactor to improve the code quality where necessary
Important exceptions were not thrown, causing the program to terminate when bugs occur.

Exceptions need to be thrown and catch to ensure the code works as intended and to eliminate bugs.

Important exceptions are thrown and catch, and the error message is displayed in Duke GUI, especially for exceptions regarding array out of bound and incorrect syntax.