Add undo feature
No undo feature to undo unwanted commands.

Having an undo feature allows users to undo an unwanted command.

Add an undo feature to allow users to undo the previous command that
edited the task list. If undo is inputted after an undo, the second undo
undoes the first undo (i.e. inputting "mark 1", "undo" and "undo" will
mark the 1st task as finished.).

Undo allows the latest command to be undone and if undo was the
unwanted command, undo can be inputted again to undo the undo command.