Improve code quality
All main logic are in Duke, resulting in poor code quality.

Lack of OOP and violation of information encapsulation. Also might be more difficult to catch errors.

Let's create separate Command classes to separate the code that is used for each one.

This makes the Duke class cleaner and demonstrates SLAP.