Add assertions
Lack of assertions.

Code makes assumptions about certain properties but this assumption is not checked. Hence could lead to errors that are undetected.

Let's add assertions to verify if crucial assumptions are indeed true.

This is such that if these assumptions are not met, we can readily spot the bug in the code and fix it before user uses it.