Not all parts of the code follow code quality standards.
ALl or most parts should follow these standards so that the code is
more readable and minimizes unsafe practices.

Some parts of the code in Duke.java is edited to make it more
readable. Unused exception classes were deleted to minimize unsafe
practices.

These were the parts of code currently identified to not be up to
code quality standards, and were hence edited accordingly.