Having notes about each task is not supported yet.
The user may want to add more notes on top of the task description.
Hence there is a need to support this system.

Implemented adding, editing and deleting of notes for each task.

This is to allow the user to add notes to each task and edit or
delete them as they wish.

User command syntax is as follows
1. Add or edit note
note /task (task number) /edit (new note)

2. Delete task
note /task (task number) /delete