Message Class: Change invalid user input message
There exists a help and an advanced help message to guide the user
on what commands can be utilised. This means that the invalid user
input message which guides the user on Duke commands is not
necessary.

Let's,
* change the content in the invalid user input message to
be more meaningful and more concise, providing a shorter
and more readable instruction for the user