Update bye command to quit app
Bye command: save tasks data and show goodbye message

Bye command does not close application.

A bye command closing the application will enhance user experience
as it is more convenient to save and close app immediately with
the same command.

Let's update the program to check if input is a bye command and
close the application immediately if so.