do C-Sort
Some users may want their tasks to be sorted in order of time.

Now users have the sort command that will sort their tasks.

Todos always come first. Deadlines and Events are sorted by their
"essential time"s. The essential time for a Deadline is the deadline
itself, while the essential time for an Event depends on whether the
event is upcoming (start time) or not (end time).